update problem 1 & 2 test programs

+++ assignment1/src/main/java/problem2/NewCar.java
@@ -4,18 +4,18 @@ public class NewCar extends Car{
 
   public NewCar(Integer manufactureYear, MakeModel makeModel, Double mSRPrice, Integer mileage) {
     super(manufactureYear, makeModel, mSRPrice, mileage);
-    if (this.mileage.intValue() <= 50) {
+    if (this.mileage <= 50) {
       Integer newCarWithinFiftyMiles = VehicleSystem.getNewCarWithinFiftyMiles();
-      VehicleSystem.setNewCarWithinFiftyMiles(Integer.valueOf(newCarWithinFiftyMiles.intValue() + 1));
+      VehicleSystem.setNewCarWithinFiftyMiles(newCarWithinFiftyMiles + 1);
     }
   }
   @Override
   public void setMileage(Integer mileage) {
     Integer newCarWithinFiftyMiles = VehicleSystem.getNewCarWithinFiftyMiles();
-    if (this.getMileage().intValue() > 50 && mileage.intValue() <= 50) {
-      VehicleSystem.setNewCarWithinFiftyMiles(Integer.valueOf(newCarWithinFiftyMiles.intValue() + 1));
-    } else if (this.getMileage().intValue() <= 50 && mileage.intValue() > 50) {
-      VehicleSystem.setNewCarWithinFiftyMiles(Integer.valueOf(newCarWithinFiftyMiles.intValue() - 1));
+    if (this.getMileage() > 50 && mileage <= 50) {
+      VehicleSystem.setNewCarWithinFiftyMiles(newCarWithinFiftyMiles + 1);
+    } else if (this.getMileage() <= 50 && mileage > 50) {
+      VehicleSystem.setNewCarWithinFiftyMiles(newCarWithinFiftyMiles - 1);
     }
     this.mileage = mileage;
   }
